Tests/LibWeb/Text/expected/css/insert-rule-keyframes-shadow-slotted-cross-sheet-invalidation.txt

opacity before keyframes insertRule: 1
opacity after keyframes insertRule: 0.25
